NMS-12424: base_attribute and unicast_prefix messages

+++ features/telemetry/protocols/bmp/adapter/src/main/java/org/opennms/netmgt/telemetry/protocols/bmp/adapter/openbmp/proto/records/BaseAttribute.java
@@ -49,8 +49,8 @@ public class BaseAttribute extends Record {
     public Integer asPathCount;
     public Long originAs;
     public InetAddress nextHop;
-    public Integer med;
-    public Integer localPref;
+    public Long med;
+    public Long localPref;
     public String aggregator;
     public String communityList;
     public String extCommunityList;
